[Bronze V] Title: 이상한 기호, Time: 104 ms, Memory: 14232 KB, Score: 100 point -BaekjoonHub

+++ Java/백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -6,12 +6,13 @@ public class Main {
     public static void main (String args[]) throws IOException{
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
+        Long a = Long.parseLong(st.nextToken());
+        Long b = Long.parseLong(st.nextToken());
+//        int b = Integer.parseInt(st.nextToken());
         System.out.println(go(a,b));
     }
 
-    private static int go(int a, int b) {
+    private static long go(long a, long b) {
         return(a+b)*(a-b);
     }
 }
